+++ framework/streams.h
@@ -68,7 +68,7 @@ extern int (read_stream)(InputStream * inp);
 extern int (peek_stream)(InputStream * inp);
 extern void (write_stream)(OutputStream * out, int b);
 extern void (write_block_stream)(OutputStream * out, const char * bytes, size_t size);
-extern int (splice_block_stream)(OutputStream * out, int fd, size_t size, off_t * offset);
+extern ssize_t (splice_block_stream)(OutputStream * out, int fd, size_t size, off_t * offset);
 
 extern void write_string(OutputStream * out, const char * str);
 extern void write_stringz(OutputStream * out, const char * str);
